klplib: ksrc: move download_supported_file to supported.py
The function download_supported_file does not concern ksrc, which should
only handle the kernel sources. Hence add a new module to retrieve the
supported codestreams.
